requirements.txt
Django==5.0.6
djangorestframework==3.15.2
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
Pillow==10.4.0

flake8==7.0.0
pytest==8.3.2
pytest-django==4.8.0